Ensure no template deduction issues with std::min

+++ include/GEA2.h
@@ -223,7 +223,7 @@ class GEA2 {
     readERDAsync(
       address, erd, reinterpret_cast<void*>(callback), +[](void* context, ReadStatus status, const void* value_, uint8_t valueSize) {
         T value;
-        memcpy(&value, value_, std::min(valueSize, sizeof(T)));
+        memcpy(&value, value_, std::min(static_cast<size_t>(valueSize), sizeof(T)));
         reinterpret_cast<void (*)(ReadStatus, T)>(context)(status, value);
       });
   }
@@ -247,7 +247,7 @@ class GEA2 {
     readERDAsync(
       address, erd, reinterpret_cast<void*>(wrappedContext), +[](void* context_, ReadStatus status, const void* value_, uint8_t valueSize) {
         T value;
-        memcpy(&value, value_, std::min(valueSize, sizeof(T)));
+        memcpy(&value, value_, std::min(static_cast<size_t>(valueSize), sizeof(T)));
         auto* wrappedContext = reinterpret_cast<WrappedContext*>(context_);
         wrappedContext->callback(wrappedContext->context, status, value);
         delete wrappedContext;
